scan: move get_patched_kernels from ksrc
